fix(api-audit-logs): skip logging private model actions
