--- packages/api-audit-logs/src/subscriptions/headlessCms/entryRevisions.ts
@@ -1,98 +1,0 @@
-import WebinyError from "@webiny/error";
-
-import { AUDIT } from "~/config";
-import { getAuditConfig } from "~/utils/getAuditConfig";
-import { isSearchModelEntry } from "./utils/isSearchModelEntry";
-import { AuditLogsContext } from "~/types";
-
-export const onEntryRevisionAfterCreateHook = (context: AuditLogsContext) => {
-    context.cms.onEntryRevisionAfterCreate.subscribe(async ({ entry }) => {
-        try {
-            if (isSearchModelEntry(entry.modelId)) {
-                return;
-            }
-
-            const createAuditLog = getAuditConfig(AUDIT.HEADLESS_CMS.ENTRY_REVISION.CREATE);
-
-            await createAuditLog("Entry revision created", entry, entry.id, context);
-        } catch (error) {
-            throw WebinyError.from(error, {
-                message: "Error while executing onEntryRevisionAfterCreateHook hook",
-                code: "AUDIT_LOGS_AFTER_ENTRY_REVISION_CREATE_HOOK"
-            });
-        }
-    });
-};
-
-export const onEntryRevisionAfterUpdateHook = (context: AuditLogsContext) => {
-    context.cms.onEntryAfterUpdate.subscribe(async ({ entry, original }) => {
-        try {
-            if (isSearchModelEntry(entry.modelId)) {
-                return;
-            }
-
-            const createAuditLog = getAuditConfig(AUDIT.HEADLESS_CMS.ENTRY_REVISION.UPDATE);
-
-            await createAuditLog(
-                "Entry revision updated",
-                { before: original, after: entry },
-                entry.id,
-                context
-            );
-        } catch (error) {
-            throw WebinyError.from(error, {
-                message: "Error while executing onEntryRevisionAfterUpdateHook hook",
-                code: "AUDIT_LOGS_AFTER_ENTRY_REVISION_UPDATE_HOOK"
-            });
-        }
-    });
-};
-
-export const onEntryRevisionAfterDeleteHook = (context: AuditLogsContext) => {
-    context.cms.onEntryRevisionAfterDelete.subscribe(async ({ entry }) => {
-        try {
-            if (isSearchModelEntry(entry.modelId)) {
-                return;
-            }
-
-            const createAuditLog = getAuditConfig(AUDIT.HEADLESS_CMS.ENTRY_REVISION.DELETE);
-
-            await createAuditLog("Entry revision deleted", entry, entry.id, context);
-        } catch (error) {
-            throw WebinyError.from(error, {
-                message: "Error while executing onEntryRevisionAfterDeleteHook hook",
-                code: "AUDIT_LOGS_AFTER_ENTRY_REVISION_DELETE_HOOK"
-            });
-        }
-    });
-};
-
-export const onEntryRevisionAfterPublishHook = (context: AuditLogsContext) => {
-    context.cms.onEntryAfterPublish.subscribe(async ({ entry }) => {
-        try {
-            const createAuditLog = getAuditConfig(AUDIT.HEADLESS_CMS.ENTRY_REVISION.PUBLISH);
-
-            await createAuditLog("Entry revision published", entry, entry.id, context);
-        } catch (error) {
-            throw WebinyError.from(error, {
-                message: "Error while executing onEntryRevisionAfterPublishHook hook",
-                code: "AUDIT_LOGS_AFTER_ENTRY_REVISION_PUBLISH_HOOK"
-            });
-        }
-    });
-};
-
-export const onEntryRevisionAfterUnpublishHook = (context: AuditLogsContext) => {
-    context.cms.onEntryAfterUnpublish.subscribe(async ({ entry }) => {
-        try {
-            const createAuditLog = getAuditConfig(AUDIT.HEADLESS_CMS.ENTRY_REVISION.UNPUBLISH);
-
-            await createAuditLog("Entry revision unpublished", entry, entry.id, context);
-        } catch (error) {
-            throw WebinyError.from(error, {
-                message: "Error while executing onEntryRevisionAfterUnpublishHook hook",
-                code: "AUDIT_LOGS_AFTER_ENTRY_REVISION_UNPUBLISH_HOOK"
-            });
-        }
-    });
-};
